<fix> 修改地区查询
修改地区查询

+++ src/main/java/qxx/information/mapper/SysRegionMapper.java
@@ -24,6 +24,6 @@ import java.util.List;
 public interface SysRegionMapper extends BaseMapper<SysRegion> {
 
 
-    List<SysRegion> queryByIdRegion(Long id);
+    List<SysRegion> queryByIdRegion(int id);
 
 }
